example/example_test.go
package example

import (
	"testing"
)

func TestListChannels(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "testListChannels"},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ListChannels(); (err != nil) != tt.wantErr {
				t.Errorf("ListChannels() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestAPIRateLimit(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "testAPIRateLimit", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := APIRateLimit(); (err != nil) != tt.wantErr {
				t.Errorf("APIRateLimit() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestShowChannel(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "testShowChannel", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ShowChannel(); (err != nil) != tt.wantErr {
				t.Errorf("ShowChannel() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestChannelPerformance(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestChannelPerformance", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ChannelPerformance(); (err != nil) != tt.wantErr {
				t.Errorf("ChannelPerformance() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestSearchTour(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestSearchTour", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := SearchTour(); (err != nil) != tt.wantErr {
				t.Errorf("SearchTour() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestShowTour(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestShowTour", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ShowTour(); (err != nil) != tt.wantErr {
				t.Errorf("ShowTour() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestListTourLocation(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestListTourLocation", wantErr: false},
	}

	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ListTourLocation(); (err != nil) != tt.wantErr {
				t.Errorf("ListTourLocation() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestListProductFilters(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestListProductFilters", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ListProductFilters(); (err != nil) != tt.wantErr {
				t.Errorf("ListProductFilters() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestShowTourDatesAndDeals(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestShowTourDatesAndDeals", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ShowTourDatesAndDeals(); (err != nil) != tt.wantErr {
				t.Errorf("ShowTourDatesAndDeals() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestUpdateTourUrl(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestUpdateTourUrl", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := UpdateTourUrl(); (err != nil) != tt.wantErr {
				t.Errorf("UpdateTourUrl() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestSearchHotelsSpecific(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestSearchHotelsSpecific", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := SearchHotelsSpecific(); (err != nil) != tt.wantErr {
				t.Errorf("SearchHotelsSpecific() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestSearchRawDepartures(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestSearchRawDepartures", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := SearchRawDepartures(); (err != nil) != tt.wantErr {
				t.Errorf("SearchRawDepartures() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestShowDeparture(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestShowDeparture", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ShowDeparture(); (err != nil) != tt.wantErr {
				t.Errorf("ShowDeparture() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestCreateDeparture(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{name: "TestCreateDeparture", wantErr: false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := CreateDeparture(); (err != nil) != tt.wantErr {
				t.Errorf("CreateDeparture() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestUpdateDeparture(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{"TestUpdateDeparture", false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := UpdateDeparture(); (err != nil) != tt.wantErr {
				t.Errorf("UpdateDeparture() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestDeleteDeparture(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{"TestDeleteDeparture", false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := DeleteDeparture(); (err != nil) != tt.wantErr {
				t.Errorf("DeleteDeparture() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestListTours(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{"TestListTours", false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ListTours(); (err != nil) != tt.wantErr {
				t.Errorf("ListTours() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestListTourImages(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{"TestListTourImages", false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ListTourImages(); (err != nil) != tt.wantErr {
				t.Errorf("ListTourImages() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}

func TestShowTourDepartures(t *testing.T) {
	tests := []struct {
		name    string
		wantErr bool
	}{
		{"TestShowTourDepartures", false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if err := ShowTourDepartures(); (err != nil) != tt.wantErr {
				t.Errorf("ShowTourDepartures() error = %v, wantErr %v", err, tt.wantErr)
			}
		})
	}
}
